Mobile Sizing Kinks
Got most of the kinks out when it comes to using mobile. The window is now the proper size, the windows will resize on window change, they windows will move to position 0 so they can resize when they detect they go over window size.

Only issue is when switching between mobile phones it can cause some small scroll issue, and sometimes the window is slightly larger than it should be.

Ideas on a complete implementation would be to use RNDs update functionality. Since clicking a window triggers the correct behavior, so using the built in update rather than manually updating position could be the final fix.